:hammer: Restructured inst/include
* The package-specific folder called "placeholder" is now called "package-code". There is no specific reason for this.
* Added a new .hpp-file which are going handle all include statements. This will be effectuated when the dependency on RcppEigen is over.

NOTE: Removed include statement and a small if-statement for testing purposes.
