+++ inst/include/SLmetrics.h
@@ -1,10 +1,9 @@
 #ifndef SLmetric_h
 #define SLmetric_h
 
-
-#include "placeholder/statistics.hpp"
-#include "placeholder/regression_task.hpp"
-#include "placeholder/templates.hpp"
+#include "package-code/statistics.hpp"
+#include "package-code/regression_task.hpp"
+#include "package-code/templates.hpp"
 #include <armadillo>
 #include <Rcpp.h>
