+++ inst/include/package-code/libraries.hpp
@@ -1,0 +1,21 @@
+/**
+ * @file libraries.hpp
+ * 
+ * @brief This file includes
+ * all relevant #include-staements 
+ * to reduce the amount of repeated code
+ *
+ * @note This will be used when
+ * <RcppEigen.h> is gone for real 
+ */
+#ifndef __libraries_hpp__
+#define __libraries_hpp__
+
+// 1) include armadille
+// and Rcpp
+// NOTE: It seems that the order matters
+#include <armadillo>
+#include <Rcpp.h>
+
+
+#endif
