+++ inst/include/package-code/templates.hpp
